requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
dulwich>=0.21.0
filelock>=3.12.0
ruamel.yaml>=0.17.0
jsonschema>=4.17.0
fastjsonschema>=2.16.0
python-ulid>=1.1.0
python-dateutil>=2.8.0
django-filter>=23.0,<25.0
